+++ test/browser/test.ml
@@ -1,1 +1,2 @@
-let () = Alcotest.run "browser" [ Test_browser.suite ]
+let () =
+  Alcotest.run "browser" [ Test_browser.suite; Test_browser_compare.suite ]